+++ src/php/Compiler/Domain/Emitter/OutputEmitter/NodeEmitter/VarEmitter.php
@@ -7,9 +7,9 @@ namespace Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter;
 use Phel\Compiler\Domain\Analyzer\Ast\AbstractNode;
 use Phel\Compiler\Domain\Analyzer\Ast\VarNode;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitterInterface;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\PhpStringEscape;
 use Phel\Lang\Registry;
 
-use function addslashes;
 use function assert;
 
 final class VarEmitter implements NodeEmitterInterface
@@ -22,9 +22,9 @@ final class VarEmitter implements NodeEmitterInterface
 
         $this->outputEmitter->emitContextPrefix($node->getEnv(), $node->getStartSourceLocation());
         $this->outputEmitter->emitStr(Registry::class . '::getInstance()->getVar("');
-        $this->outputEmitter->emitStr(addslashes($this->outputEmitter->mungeEncodeRegistryKey($node->getNamespace())));
+        $this->outputEmitter->emitStr(PhpStringEscape::doubleQuoted($this->outputEmitter->mungeEncodeRegistryKey($node->getNamespace())));
         $this->outputEmitter->emitStr('", "');
-        $this->outputEmitter->emitStr(addslashes($node->getName()));
+        $this->outputEmitter->emitStr(PhpStringEscape::doubleQuoted($node->getName()));
         $this->outputEmitter->emitStr('")');
         $this->outputEmitter->emitContextSuffix($node->getEnv(), $node->getStartSourceLocation());
     }